srcAI/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
requests==2.32.5
pydantic==2.12.5
sentence-transformers==5.2.0
chromadb==1.3.6
numpy==2.3.5
scikit-learn==1.8.0
scipy==1.16.3
torch==2.9.1
fpdf2==2.8.1
